refactor: restructured project to match the standard golang project structure
